+++ src/scene/camera.hpp
@@ -11,21 +11,21 @@ class Camera {
     Vec3 front;
     Vec3 right;
     Point3 lower_left_corner;
-    double viewport_h;
-    double viewport_v;
-    double fov_deg;
-    double aspect_ratio;
-    double aperture;
-    double focus_dist;
+    float viewport_h;
+    float viewport_v;
+    float fov_deg;
+    float aspect_ratio;
+    float aperture;
+    float focus_dist;
 public:
     Camera(
         const Point3& origin,
         const Point3& lookat,
         const Vec3& view_up,
-        double fov_deg,
-        double aspect_ratio,
-        double aperture,
-        double focus_dist
+        float fov_deg,
+        float aspect_ratio,
+        float aperture,
+        float focus_dist
     ):
     origin(origin), 
     lookat(lookat),  
@@ -36,13 +36,13 @@ public:
         front = unit_vector(lookat - origin);
         right = unit_vector(cross(front, view_up));
         up = cross(right, front);
-        double fov_rad = fov_deg * std::numbers::pi / 180.0;
-        viewport_v = 2.0 * tan(fov_rad * 0.5) * focus_dist;
+        float fov_rad = fov_deg * static_cast<float>(std::numbers::pi) / 180.0f;
+        viewport_v = 2.0f * tan(fov_rad * 0.5f) * focus_dist;
         viewport_h = aspect_ratio * viewport_v;
-        lower_left_corner = origin - viewport_h * 0.5 * right - viewport_v * 0.5 * up  + front * focus_dist;
+        lower_left_corner = origin - viewport_h * 0.5f * right - viewport_v * 0.5f * up  + front * focus_dist;
     }
 
-    Ray get_ray(double u, double v) const {
+    Ray get_ray(float u, float v) const {
         return Ray(origin, lower_left_corner + u * viewport_h * right + v * viewport_v * up - origin);
     }
 };
